Add utility functions for text and block handling; implement alignment, rotation, and scaling features
