+++ commands/barCodeCommand.js
@@ -1,0 +1,69 @@
+import {getScale, getUnit} from '../store/state.js';
+
+export function handleBarcodeCommand(line, previewDiv) {
+  const SCALE = getScale();
+  const UNIT = getUnit();  
+  
+  const barcodeRegex =
+      /BARCODE\s+(\d+)\s*,\s*(\d+)\s*,\s*"([^"]+)"\s*,\s*(\d+)\s*,\s*(\d+)\s*,\s*(\d+)\s*,\s*(\d+)\s*,\s*(\d+)(?:\s*,\s*(\d+))?\s*,\s*"([^"]+)"\s*/i;
+    const match = line.match(barcodeRegex);
+    if (match) {
+      const x = parseInt(match[1]);
+      const y = parseInt(match[2]);
+      const codeType = match[3];
+      const heightBarcode = parseInt(match[4]);
+      const humanReadable = parseInt(match[5]);
+      const rotation = parseInt(match[6]);
+      const narrow = parseInt(match[7]);
+      const wide = parseInt(match[8]);
+  
+      let alignment, content;
+      if (match[9] !== undefined && match[10] !== undefined) {
+        alignment = parseInt(match[9]);
+        content = match[10];
+      } else {
+        alignment = humanReadable;
+        content = match[10];
+      }
+  
+      const barcodeContainer = document.createElement("div");
+      barcodeContainer.style.position = "absolute";
+      barcodeContainer.style.left = x * SCALE + UNIT;
+      barcodeContainer.style.top = y * SCALE + UNIT;
+  
+      // Se utiliza una fórmula simple: (narrow+wide)* (longitud del contenido)* factor base (8 dots)
+      let barcodeWidth = (narrow + wide) * content.length * 6;
+      // Aplicamos la escala
+      barcodeWidth = barcodeWidth * SCALE;
+  
+      const barcodeBar = document.createElement("div");
+      barcodeBar.style.width = barcodeWidth + UNIT;
+      barcodeBar.style.height = heightBarcode * SCALE + UNIT;
+      barcodeBar.style.backgroundColor = "black";
+      barcodeContainer.appendChild(barcodeBar);
+  
+      if (humanReadable !== 0) {
+        const textDiv = document.createElement("div");
+        textDiv.textContent = content;
+        textDiv.style.width = barcodeWidth + UNIT;
+        textDiv.style.fontSize = 30 * SCALE + UNIT; // tamaño fijo para previsualizar
+        textDiv.style.fontFamily = "monospace";
+        if (alignment === 2) {
+          textDiv.style.textAlign = "center";
+        } else if (alignment === 3) {
+          textDiv.style.textAlign = "right";
+        } else {
+          textDiv.style.textAlign = "left";
+        }
+        textDiv.style.marginTop = 2 * SCALE + UNIT;
+        barcodeContainer.appendChild(textDiv);
+      }
+  
+      if (rotation !== 0) {
+        barcodeContainer.style.transform = `rotate(${rotation}deg)`;
+        barcodeContainer.style.transformOrigin = "left";
+      }
+  
+      previewDiv.appendChild(barcodeContainer);
+    }
+  }
