+++ index.html
@@ -45,6 +45,6 @@
       </div>
     </div>
   </div>
-  <script src="script.js"></script>
+  <script src="script.js" type="module"></script>
 </body>
 </html>
